Fixed broken code inside lists in "Working with Gettit"
Markdown has a known limitation where indented blocks inside a list are
interpreted as the part of the list, instead of a code block. Breaking
beforehand with an un-indented line (even if it has nothing but &nbsp;)
fixes this issue.

This patch completes the work of the previous PR and closes issue #91

Change-Id: I376efb63b7b8b25ad61b88baa51a7431e0197d61
Signed-off-by: Allon Mureinik <amureini@redhat.com>