Add engine API explorer
This patch adds the HTML and JavaScript code of the engine API explorer
application, and the JSON file containing the description of version
4.0.20 of the API specification. The application should be available in
the following URL:

  http://www.ovirt.org/documentation/api-explorer

That should be identical to what is currently available here:

  https://jhernand.fedorapeople.org/ovirt-api-explorer

Note that the "index.js" file isn't the original source code of the
application, but the result of running "webpack" to pack together all
the JavaScript source code. The original source code is available in the
"ovirt-engine-api-explorer" project hosted in "gerrit.ovirt.org".

The "model.json" has also been automatically generated from the API
model available in the "ovirt-engine-api-model" project, version 4.0.20.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>