Remove "amurei" from authors
This user was created by accident back in the MediaWiki days. The
correct user is amureini which appears after amurei in the
authors.yml file. This is also evident by the fact that both users have
the same details (name, email and gravatar), and the email is in fact
amureini@redhat.com.

Change-Id: If4c7dad01b4d92602c6bb47ba86cb720da65a000
Signed-off-by: Allon Mureinik <amureini@redhat.com>