use code block for git clone snippets
Unlike the rest of code snippets in Vdsm Developers document, `git clone` commands are shown as inline code, not a code block. Because of that they are somehow easy to look over. For the sake of consistency and readability I change two `git clone` snippets from inline code into a code block.