identify broken links and add newer links
I notice quite a few links in the documentation lead to a "Not found" page with a set of search results.
I was quite frustrated by this. Such frustration reflects badly on a oVirt, a project that I wish to succeed.
Since I don't know a good way to find what those links should be, 
I thought it might save other users the trouble of discovering that those links are broken.