Move "Writing a new user query" to the developer guide
The "Writing a new user query" page should be part of the developer's
guide, not a feature. This patch moves it to its rightful place and
fixes the links pointing to it.

As a side "bonus" of moving it to the right place, the directory
source/develop/release-management/features/infra/user-portal-permissions
was removed, which solves middleman's inability to show the
source/develop/release-management/features/infra/user-portal-permissions.html.md
file due to naming conflicts.

Signed-off-by: Allon Mureinik <amureini@redhat.com>