Replace DC 4.1 page with Storage Domain V4 page
Each oVirt release introduces a new DC level with, potentially, many
features. This page listed a very small subset of them, which was
incomplete, and would it be completed it should not have resided under
storage.

This patch removes the aforementioned patch and replaces it with a page
discussing the actual new feature here from the storage perspective -
the V4 domain, and the features it enables (which is not, of course, a
full list of storage features in oVirt 4.1).

Signed-off-by: Allon Mureinik <amureini@redhat.com>