Update migrate-to-hosted-engine.html.md
@sandrobonazzola  I corrected a couple of links on this page. 
Is the link at the bottom of the page correct?
That's it. If we now point a browser at https://my-engine.home.local/ovirt-engine/ , we can login to the web admin interface, and in addition to our existing data, we'll find also a host called 'hosted_engine_1' and a VM called 'HostedEngine'.