SSL for install instructions
SSL everywhere is a good idea overall, but is a MUST condition for installation instructions to use sudo to install packages.

While we're at it, might as well update all the links to good ol' resources.ovirt.org .

I recommend updating the other links for the mirrors, but did not find a viable HTTPS endpoint (I didn't look further than to attempt to load the page as https://). The mirrors that did have a valid HTTPS endpoint were updated in this PR. Perhaps you can request your mirror providers to use an SSL certificate now that e.g. Let's Encrypt is ubiquitous, useful, automatable, etc.