Update IPv6 OVN subnet feature page supporting RAs
Add support for stateless DHCPv6 address configuration.

The IPv6 options are presented in a matrix, which is useful for
QEs to know what to test in which configuration, and also for
potencial users to evaluate which address mode better suits
their subnets.

Clear instructions on how to configure the system to support
both address mode types, featuring REST and ansible examples
of the required configuration.

The possible use cases are presented to the user, and their
limitations exposed, thus helping the user to select what
address mode is better suited to the requirements.