Normalize markdown file extension
This is useful for the Jekyll migration

Also fixes some links which used this old extension; nevertheless not
all are really fixed because some filenames have dots in the name
which do not separate extensions and MM is lost (like
appe-Upgrading_to_oVirt_Engine_4.2_with_ovirt-fast-forward-upgrade.md).
This will be fixed when switching to Jekyll.