Add files via upload
To promote the upcoming oVirt event in Rome here's an example of the Eventbrite image as a  banner for ovirt.org. The banner image should be evenly placed between the navigation and the home message in the top banner.  

![oVirt Rome 2019 website banner](https://user-images.githubusercontent.com/20227374/64882812-d9f28c00-d62b-11e9-85b0-5c910be4605a.jpg)

Here's the image file: 
![https___cdn evbuc com_images_72213595_278319565097_1_original](https://user-images.githubusercontent.com/20227374/64882813-dbbc4f80-d62b-11e9-9978-d58fce6c7135.jpeg)

The image file should link to the event page where people can sign up for it: https://www.eventbrite.it/e/biglietti-ovirt-conference-2019-innovate-the-datacenter-with-open-virtualization-71510039453#




@sandrobonazzola