incremental-backup.md: Add documentation for new backup_mode attribute
In order to fix https://bugzilla.redhat.com/1861674, a new 'backup_mode'
attribute should be added for each disk that participates in an incremental
backup. The 'backup_mode' attribute will indicate the type of backup that
was taken for each disk:

  - 'full': If a full backup was taken for the disk.
  - 'incremental': If an incremental backup was taken for the disk.

This new attribute allows the engine to take a mixed backup that includes
incremental for part of the disks and full backup for others. The mixed
backup is needed in case that the user includes disks that weren't part
of the given checkpoint ID (from_checkpoint_id), a full backup should be
taken for those disks while an incremental backup will be taken for the
other disks.

Change-Id: I57fa40a9c7087d97fcacfdcc8ff40b8f3bdf466c
Bug-Url: https://bugzilla.redhat.com/1861674
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>