Update tox/venv devel setup guide
The devel guide instructs developers to
install tox directly on their system.
This is however not needed and moreover
could lead to installing wrong version
of tox (and other dependencies).

Update this part of the guide to point
developers to README file in tree root,
which describes VDSM venv setup.

That way it is expected that everyone
will use packages we define in
requirements.txt