Switch to GH actions and test on prod system
The goal is to be more integrated. The CI currently is associated with a
specific user account on travis-ci.org but it's better for the team to
have more control over it (or in case the person leaves the project).

I also made changes to test the website on the same OS used by the
production builder. This way changes to the Gemfile can be tested by the
team directly.