Added Planning and Planning_and_Prerequisites_Guide
This guide is modelled on the doc- original content, but has been updated to refer to oVirt and removed references to Red Hat content (which may disappear in the future).

The material hasn't been updated to reflect the latest oVirt content, but is a starting point for further updates.

Red Hat links were left with placeholders for now.

Signed-off-by: Geoff O'Callaghan <geoffocallaghan@gmail.com>