ui: fix hyperlinks to the mailing list
Due to the needed move of the mailing list the hyperlinks were broken.
For most of the mentioned archived posts a new hyperlink was able to be found.
Some were removed due to it being impossible to trace it back which mail thread it was linked to.
Fixes [issue](https://github.com/oVirt/ovirt-site/issues/3211).

Signed-off-by: Jasper Berton <jasper.berton@team.blue>