Drop events from community
In the past it was common to have oVirt events along with other Red Hat
OSAS coordinated events. It has been already several years since last
time this happened and removing the dependencies on OSAS calendar allows
to drop ruby deps making it a bit easier to try moving forward to a new
jekyl version.

Signed-off-by: Sandro Bonazzola <sandro.bonazzola@gmail.com>